requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
